refactor: 폴더 구조 개선
- google maps api 관련 컴포넌트 이동


[#61]
